% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{outlierdetection}
\alias{outlierdetection}
\title{Helper for outlier detection}
\usage{
outlierdetection(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
5-number summary, iqr, lower fence and upper fence
}
\description{
Computes 5-number summary by using
   exclusive method of quartile calculation.
   Also outputs interquartile range and lower
   and upper fences to detect outliers.
   Prints outliers, if there are any.
}
\examples{
outlierdetection(mtcars$wt)
}
