% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_z_test.R
\name{pwr.z.test}
\alias{pwr.z.test}
\title{Power of a Z-Test for Population Mean}
\usage{
pwr.z.test(
  m0,
  m1,
  sd,
  n,
  sig.level = 0.05,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{m0}{The value of the population mean from the null hypothesis}

\item{m1}{An alternative value of the population mean}

\item{sd}{Population standard deviation}

\item{n}{Sample size}

\item{sig.level}{Significance level, default is 0.05}

\item{alternative}{type of alternative, with values
"two.sided","less" or "greater"}
}
\value{
Power of the test
}
\description{
Computes the Power of a Z-test for the Population Mean
}
\examples{
pwr.z.test(m0 = 3, m1 = 2.5, sd = 1, n = 15,
sig.level = 0.01, alternative = "two.sided")
pwr.z.test(m0 = 3, m1 = 3.5, sd = 1, n = 20,
sig.level = 0.10, alternative = "greater")
}
