% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssize_p_test.R
\name{ssize.p.test}
\alias{ssize.p.test}
\title{Sample Size of a Z-Test for Population Proportion}
\usage{
ssize.p.test(p0, p1, power = 0.9, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{p0}{The value of the population proportion from the
null hypothesis (in decimal form)}

\item{p1}{An alternative value of the population proportion}

\item{power}{Power of the test}

\item{sig.level}{Significance level, default is 0.05}

\item{alternative}{type of alternative, with values
"two.sided","less" or "greater"}
}
\value{
Sample size required to achieve the given power
}
\description{
Computes the Sample Size of a Z-test for the
Population Proportion
}
\examples{
ssize.p.test(p0 = 0.10, p1 = 0.08, power = 0.9,
sig.level = 0.01, alternative = "two.sided")
ssize.p.test(p0 = 0.30, p1 = 0.25, power = 0.92,
sig.level = 0.10, alternative = "less")
}
