% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssize_z_test.R
\name{ssize.z.test}
\alias{ssize.z.test}
\title{Power of a Z-Test for Population Mean}
\usage{
ssize.z.test(
  m0,
  m1,
  sd,
  power = 0.9,
  sig.level = 0.05,
  alternative = "two.sided"
)
}
\arguments{
\item{m0}{The value of the population mean from the null hypothesis}

\item{m1}{An alternative value of the population mean}

\item{sd}{Population standard deviation}

\item{power}{Power of the test}

\item{sig.level}{Significance level, default is 0.05}

\item{alternative}{type of alternative, with values
"two.sided","less" or "greater"}
}
\value{
Sample size required to achieve the given power
}
\description{
Computes the Power of a Z-test for the Population Mean
}
\examples{
ssize.z.test(m0 = 3.25, m1 = 3.5, sd = 1, power = 0.9,
sig.level = 0.01, alternative = "two.sided")
ssize.z.test(m0 = 3.25, m1 = 3, sd = 1, power = 0.92,
sig.level = 0.10, alternative = "less")
}
